Test: commit pre-adding jquery by yarn
