+++ .gitignore
@@ -29,3 +29,7 @@
 
 /config/master.key
 coverage 
+
+/db/schema.rb
+
+entrypoint.sh
