+++ app/assets/javascripts/marker_list.js
