+++ app/views/authorities/index.html.erb
@@ -1,3 +1,4 @@
+<%= javascript_include_tag "auth_loc" %>
 <h1> Página de Gerenciamento de Desastres</h1>
 
 <p><%= link_to "Log out", '/authorities/sign_out', method: :delete %></p>
