+++ app/views/homes/index.html.erb
@@ -11,5 +11,4 @@
     <% end %>
     </div>
 <div id="map"></div>
-
 <%= javascript_include_tag "homes" %>
