+++ app/views/markers/list_disasters.html.erb
@@ -1,4 +1,4 @@
-<%= javascript_include_tag "auth_list" %>
+<%= javascript_include_tag "marker_list" %>
 
 <h1>Disaster List</h1>
