+++ config/initializers/assets.rb
@@ -12,4 +12,4 @@ Rails.application.config.assets.paths << Rails.root.join('node_modules')
 # application.js, application.css, and all non-JS/CSS in the app/assets
 # folder are already added.
 # Rails.application.config.assets.precompile += %w( admin.js admin.css )
-Rails.application.config.assets.precompile += %w(homes.js markers.js auth_list.js)
+Rails.application.config.assets.precompile += %w(homes.js markers.js marker_list.js auth_loc.js)
