+++ config/routes.rb
@@ -5,7 +5,11 @@ Rails.application.routes.draw do
   get "index/user" => "homes#user"
   get    'login'   => 'sessions#new'
   post   'login'   => 'sessions#create'
-  delete 'sign_out'  => 'sessions#destroy'
+  delete 'sign_out'  => 'sessions#destroy'  
+  post "setLocation" => 'sessions#setLocation'
+
+  
+  get "markers/list" => 'markers#list_disasters'
   
   get '/users/profile' => 'users#show'
   get '/users/edit' => 'users#edit'
@@ -17,6 +21,4 @@ Rails.application.routes.draw do
   post "authorities/login" => 'authorities#create'
   get "authorities" => 'authorities#index'
   delete '/authorities/sign_out' => 'authorities#destroy'
-  get "authorities/list" => 'authorities#list_disasters'
-  get "authorities/setLocation" => 'authorities#setLocation'
 end
