+++ spec/helpers/sessions_helper_spec.rb
@@ -1,0 +1,12 @@
+require 'rails_helper'
+
+RSpec.describe SessionsHelper, type: helper do
+   # it 'Mesmo lugar' do
+#        pending
+        #helper.calculateDistance('0','0','0','0')
+    #end
+
+    #it 'Calcula distancia' do
+     #   helper.calculateDistance('26.1232','-23.3323','-23.4823919','-46.5004498').to eq(6052.83)
+    #end
+end
